# DTOs/AttendanceDtos.cs
namespace LabManagementBackend.DTOs
{
    public class ClockInDto
    {
        public string LabId { get; set; }
    }

    public class ClockOutDto
    {
        public string LabId { get; set; }
    }
}